+++ server/src/main/java/org/apache/iotdb/db/engine/compaction/reader/PointPriorityReader.java
@@ -54,7 +54,7 @@ public class PointPriorityReader {
         new PriorityQueue<>(
             (o1, o2) -> {
               int timeCompare = Long.compare(o1.timestamp, o2.timestamp);
-              return timeCompare != 0 ? timeCompare : Integer.compare(o2.priority, o1.priority);
+              return timeCompare != 0 ? timeCompare : Long.compare(o2.priority, o1.priority);
             });
   }
